move out navigation into content script
