+++ src/handlers/navigation-handler.ts
@@ -1,5 +1,3 @@
-import CommandHandler from "../command-handler";
-
 /*
  * Handles commands from the client about navigation.
  */
@@ -9,11 +7,11 @@ export default class NavigationHandler {
   postMessage?: (request: string, data?: any) => Promise<any>;
 
   async COMMAND_TYPE_BACK(_data: any): Promise<any> {
-    CommandHandler.executeScript("window.history.back();");
+    return this.postMessage!("navigateBack");
   }
 
   async COMMAND_TYPE_FORWARD(_data: any): Promise<any> {
-    CommandHandler.executeScript("window.history.forward();");
+    return this.postMessage!("navigateForward");
   }
 
   async COMMAND_TYPE_RELOAD(_data: any): Promise<any> {
@@ -22,23 +20,8 @@ export default class NavigationHandler {
 
   async COMMAND_TYPE_SCROLL(data: any): Promise<any> {
     // Scrolling in a direction
-    let direction;
-    switch (data.direction) {
-      case "left":
-        direction = `left: -window.innerWidth * 0.8`;
-        break;
-      case "right":
-        direction = `left: window.innerWidth * 0.8`;
-        break;
-      case "up":
-        direction = `top: -window.innerWidth * 0.8`;
-        break;
-      case "down":
-        direction = `top: window.innerWidth * 0.8`;
-        break;
-    }
-    if (direction) {
-      CommandHandler.executeScript(`window.scrollBy({ ${direction}, behavior: "smooth" });`);
+    if (data.direction) {
+      return this.postMessage!("scrollDirection", { direction: data.direction });
     }
 
     // Scrolling to a path
